az_func/ph_check_function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
phonenumbers==8.12.15